+++ packages/react/src/modules/ms-teams/components/MsTeamsChannelCombobox/MsTeamsChannelInTeamCombobox.tsx
@@ -107,11 +107,15 @@ export const MsTeamsChannelInTeamCombobox: FunctionComponent<
           }
           closeOnSelect={false}
           layout="wrap"
+          modal={
+            // Modal comboboxes cause page layout to shift when body has padding. See KNO-7854.
+            false
+          }
         >
           <Combobox.Trigger />
           <Combobox.Content>
             <Combobox.Search className="rtk-combobox__search" />
-            <Combobox.Options className="rtk-combobox__options">
+            <Combobox.Options maxHeight="36">
               {sortedChannels.map((channel) => (
                 <Combobox.Option key={channel.id} value={channel.id}>
                   {channel.displayName}