+++ packages/react/src/modules/slack/components/SlackChannelCombobox/icons/index.ts
@@ -1,5 +1,0 @@
-export * from "./CheckmarkIcon";
-export * from "./InfoIcon";
-export * from "./HashtagIcon";
-export * from "./LockIcon";
-export * from "./SearchIcon";