+++ packages/expo/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @knocklabs/expo
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [7904b65]
+  - @knocklabs/react-core@0.4.0
+  - @knocklabs/react-native@0.5.10
+
+## 0.2.9
+
+### Patch Changes
+
+- Updated dependencies [8ea25f4]
+  - @knocklabs/client@0.11.4
+  - @knocklabs/react-core@0.3.4
+  - @knocklabs/react-native@0.5.9
+
 ## 0.2.8
 
 ### Patch Changes